// front/src/types/api.ts
export type IUpdateUser = {
  update: boolean;
  userId: string;
  surname: string;
  middlename: string;
  username: string;
  email?: string;
  birthday: number;
  avatar: string;
};